fixes with unauthorized
