+++ skypro-music/src/api/tracksApi.ts
@@ -1,5 +1,5 @@
 import { Track } from "@/types/Track.types";
-import { fetchWithAuth } from "./userApi";
+import { fetchWithAuth } from "@/utils/fetchWithAuth";
 
 const baseHost = "https://webdev-music-003b5b991590.herokuapp.com";
 
@@ -19,16 +19,14 @@ export async function getAllTracks(): Promise<Array<Track>> {
 export type FavoriteRequestProps = {
   id: number | undefined;
   access: string;
-  refresh: string;
 };
 
 // Добавить трек в избранное по id
 export async function addFavorite({
   id,
   access,
-  refresh,
 }: FavoriteRequestProps) {
-  const response = await fetchWithAuth(
+  const response = await fetch(
     baseHost + "/catalog/track/" + id + "/favorite/",
     {
       method: "POST",
@@ -36,7 +34,6 @@ export async function addFavorite({
         Authorization: `Bearer ${access}`,
       },
     },
-    refresh
   );
 
   if (response.status === 401) {
@@ -54,9 +51,8 @@ export async function addFavorite({
 export async function removeFavorite({
   id,
   access,
-  refresh,
 }: FavoriteRequestProps) {
-  const response = await fetchWithAuth(
+  const response = await fetch(
     baseHost + "/catalog/track/" + id + "/favorite/",
     {
       method: "DELETE",
@@ -64,7 +60,6 @@ export async function removeFavorite({
         Authorization: `Bearer ${access}`,
       },
     },
-    refresh
   );
 
   if (response.status === 401) {
