+++ skypro-music/src/api/userApi.ts
@@ -21,18 +21,17 @@ export async function register({ email, password, username }: RegisterProps) {
     },
   });
 
-  if (response.status === 412) {
-    throw new Error("Данные в неверном формате");
-  }
   if (response.status === 403) {
     throw new Error("Введенный Email уже занят");
+  }
+  if (response.status === 412) {
+    throw new Error("Данные в неверном формате");
   }
   if (response.status === 500) {
     throw new Error("Ошибка при получении данных");
   }
 
   const data = await response.json();
-
   return data.data;
 }
 
@@ -46,18 +45,20 @@ export async function login({ email, password }: LoginProps) {
     },
   });
 
-  if (response.status === 412) {
-    throw new Error("Данные в неверном формате");
+  if (response.status === 400) {
+    throw new Error("Запрос составлен некорректно");
   }
   if (response.status === 401) {
     throw new Error("Пользователь с таким email или паролем не найден");
+  }
+  if (response.status === 412) {
+    throw new Error("Данные в неверном формате");
   }
   if (response.status === 500) {
     throw new Error("Ошибка при получении данных");
   }
 
   const data = await response.json();
-
   return data;
 }
 
@@ -65,33 +66,6 @@ export type GetTokenProps = {
   email: string;
   password: string;
 };
-
-export async function fetchWithAuth(
-  url: string,
-  options: {
-    method: string,
-    headers: HeadersInit;
-  },
-  refresh: string
-) {
-  let response = await fetch(url, options);
-
-  if (response.status === 401) {
-    const newAccessToken = await refreshToken(refresh);
-
-    options.headers = {
-      ...options.headers,
-      Authorization: `Bearer ${newAccessToken}`,
-    };
-    response = await fetch(url, options);
-  }
-
-  if (response.status === 500) {
-    throw new Error(response.statusText);
-  }
-
-  return response;
-}
 
 // Получить токен
 export async function getToken({ email, password }: GetTokenProps) {
@@ -117,26 +91,4 @@ export async function getToken({ email, password }: GetTokenProps) {
   return data;
 }
 
-// Обновить токен
-export async function refreshToken(refresh: string) {
-  const response = await fetch(baseHost + "/user/token/refresh/", {
-    method: "POST",
-    body: JSON.stringify(refresh),
-    headers: {
-      "content-type": "application/json",
-    },
-  });
 
-  if (response.status === 400) {
-    throw new Error("Токен недействителен или просрочен");
-  }
-  if (response.status === 401) {
-    throw new Error("В теле запроса не передан refresh токен");
-  }
-  if (response.status === 500) {
-    throw new Error("Ошибка при получении данных");
-  }
-
-  const data = await response.json();
-  return data;
-}
